Squashed commit of Coffeescript support!
First experience with Coffeescript. This app doesn't have much JS
yet, but it seemed like a good excuse to learn Coffeescript, so
there you go. The conversion saved a few lines of closing curly-braces
and cleaned up the code a wee bit.

N.b. that we've also stopp versioning Javascript files now since
they're now generated by the Coffeescript compiler.

Finally, and somewhat oddly, we compile coffeescript into ./tmp/
to account for the fact that, on Heroku's R/O fs, we can't save
files into public/javascripts. Theoretically on Heroku's new Cedar
stack this is no longer a problem, since the "Ephemeral Filesystem"
should be writable [1], but in interest of portability, we're erring
on the safe side by not requiring a writable ./public/.

[1] http://devcenter.heroku.com/articles/dyno-isolation#ephemeral_filesystem